+++ src/elements/content-sidebar/activity-feed/task-new/TaskDueDate.js
@@ -10,8 +10,8 @@ import { TASK_NEW_NOT_STARTED } from '../../../../constants';
 
 import './TaskDueDate.scss';
 
-import type { TaskStatus } from './flowTypes';
-import type { ISODate } from '../../../../common/flowTypes';
+import type { TaskStatus } from '../../../../common/types/tasks';
+import type { ISODate } from '../../../../common/types/core';
 
 type Props = {
     dueDate: ISODate,
