+++ src/features/unified-share-modal/InviteePermissionsDescription.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 
-import type { ItemType } from '../../common/flowTypes';
+import type { ItemType } from '../../common/types/core';
 
 import { EDITOR, CO_OWNER, PREVIEWER, PREVIEWER_UPLOADER, VIEWER, VIEWER_UPLOADER, UPLOADER } from './constants';
 import messages from './messages';
