+++ src/features/unified-share-modal/SharedLinkSection.js
@@ -17,7 +17,7 @@ import IconMail from '../../icons/general/IconMail';
 import IconClock from '../../icons/general/IconClock';
 import IconGlobe from '../../icons/general/IconGlobe';
 import { amaranth } from '../../styles/variables';
-import type { ItemType } from '../../common/flowTypes';
+import type { ItemType } from '../../common/types/core';
 import { isBoxNote } from '../../utils/file';
 
 import convertToBoxItem from './utils/item';
